threads per cpu supports 1/2 and 1/4 for lighter load

+++ fraktal_sft/Settings.h
@@ -48,7 +48,7 @@ private:
   int m_WindowRight;
   int m_ImageWidth;
   int m_ImageHeight;
-  int m_ThreadsPerCore;
+  double m_ThreadsPerCore;
   bool m_AnimateZoom;
   bool m_ArbitrarySize;
   bool m_ReuseReference;
@@ -164,8 +164,8 @@ public:
   inline int    GetImageHeight() const { return m_ImageHeight; };
   inline void   SetImageHeight(int h) { m_ImageHeight = h; };
 
-  inline int    GetThreadsPerCore() const { return m_ThreadsPerCore; };
-  inline void   SetThreadsPerCore(int t) { m_ThreadsPerCore = t; };
+  inline double GetThreadsPerCore() const { return m_ThreadsPerCore; };
+  inline void   SetThreadsPerCore(double t) { m_ThreadsPerCore = t; };
 
   inline bool   GetAnimateZoom() const { return m_AnimateZoom; };
   inline void   SetAnimateZoom(bool b) { m_AnimateZoom = b; };
